+++ angular/projects/wcs-angular/src/lib/directives/grid-value-accessor.directive.ts
@@ -45,4 +45,11 @@ export class GridValueAccessorDirective implements ControlValueAccessor {
   registerOnTouched(fn: any): void {
     this.onTouched = fn;
   }
+
+  @HostListener('wcsBlur', ['$event.target'])
+  _handleBlurEvent(el: any): void {
+    if (el === this.el.nativeElement) {
+      this.onTouched();
+    }
+  }
 }